+++ TestApplications/SignalR_Testapp/SignalR_Testapp/Models/PrettyItem.cs
@@ -9,7 +9,7 @@ namespace SignalR_Testapp.Models
     {
         private long _itemno;
         private string _name;
-        private long _price;
+        private int _price;
         private long _bid;
         private DateTime _expires;
         private string _description;
@@ -18,7 +18,7 @@ namespace SignalR_Testapp.Models
 
         public long itemno { get { return this._itemno; } set { this._itemno = value; } }
         public string name { get { return this._name; } set { this._name = value; } }
-        public long price { get { return this._price; } set { this._price = value; } }
+        public int price { get { return this._price; } set { this._price = value; } }
         public long bid { get { return this._bid; } set { this._bid = value; } }
         public DateTime expires { get { return this._expires; } set { this._expires = value; } }
         public string description { get { return this._description; } set { this._description = value; } }
